umpf: fix ignoring flags
The flags implementation loses the flags if they are provided via umpf
build followed by an umpf tag, e.g:

 - umpf build <useries containing flags> && umpf tag
 - umpf build <useries> --flags <flags> && umpf tag

In this case the flags are lost. Fix this by storing the flags and
adding a build note like it is done for base and name. This also fixes
the issue losing the flags while an 'umpf continue' if they are provided
via an umpf-series file.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>